zavrsen dizajn za room formu. Implementirana funkcionalnost svih dugmica kao i prikaz tabele na gridu. Povezana room forma sa main formom preko dugmeta. Sledi kreiranja forme i klase za recepciju

+++ GuestClass.cs
@@ -15,7 +15,7 @@ namespace Hotel
         {
             string insertQuerry = "INSERT INTO `guest`(`GuestID`, `GuestFirstName`, `GuestLastName`, `GuestPhone`, `GuestCity`) VALUES(@id, @fname, @lname, @phone, @city)";
             MySqlCommand command = new MySqlCommand(insertQuerry, connect.GetConnection());
-            command.Parameters.Add("@id", MySqlDbType.Int16).Value = id;
+            command.Parameters.Add("@id", MySqlDbType.Int32).Value = id;
             command.Parameters.Add("@fname", MySqlDbType.VarChar).Value = fname;
             command.Parameters.Add("@lname", MySqlDbType.VarChar).Value = lname;
             command.Parameters.Add("@phone", MySqlDbType.VarChar).Value = phone;
